Feedback items display as a list on solution show page
